Fix chalk menu scroll behaviour
This commit also changes the status bar color to light gray, whilst also
redefining the color light gray as it was swapped with dark gray.

The change of the status bar color is for letting the background be
visible on chalk so you can see that the time is missing there.
Also I think it looks rather nice aswell.